requirements.txt
GitPython
python-dotenv
langchain
langchain-openai
langchain-google-genai
